% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rcpp_hello_world}
\alias{rcpp_hello_world}
\title{Example function of rcpp}
\usage{
rcpp_hello_world(a, b)
}
\arguments{
\item{string}{first string}

\item{string}{second string}
}
\value{
list list of two vectors
}
\description{
Example function of rcpp
}

